# docs/swagger.json
{
    "produces": [
        "application/json"
    ],
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "T-Loyal API",
        "title": "Crazy Peppers T-Loyal API",
        "contact": {},
        "version": "1.0"
    },
    "host": "prod-team-29-4254c2ee.REDACTED",
    "basePath": "/api/v1",
    "paths": {
...
        "/partners/{partnerID}/stat": {
            "get": {
                "description": "Возвращает метрики использования промо-кодов, такие как активные пользователи, количество покупок, retention и т. д.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Statistics"
                ],
                "summary": "Получение статистики по промо-акциям",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID партнера",
                        "name": "partnerID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешный ответ со статистикой",
                        "schema": {
                            "$ref": "#/definitions/models.Stat"
                        }
                    },
                    "400": {
                        "description": "Ошибка валидации данных. Пример: {\\\"error\\\": \\\"Некорректный\tUUID\tпартнера\\\"}",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Ошибка авторизации. Пример: {\\\"error\\\": \\\"Некорректный\t\t\tтокен\tили\t\t\tпользователь\tне\tаутентифицирован\\\"}",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Ошибка доступа. Пример: {\\\"error\\\": \\\"Недостаточно\t\t\t\tправ\tдля\t\t\tпросмотра\t\tстатистики\\\"}",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера. Пример: {\\\"error\\\": \\\"Ошибка\t\tпри\t\tполучении\tстатистики\\\"}",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/partners/{partnerID}/stat/export": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Генерирует Excel-файл с активациями промокодов партнера",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet"
                ],
                "tags": [
                    "Statistics"
                ],
                "summary": "Экспорт статистики партнера",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID партнера",
                        "name": "partnerID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "400": {
                        "description": "Некорректный UUID",
                        "schema": {
                            "$ref": "#/definitions/utils.Err"
                        }
                    },
                    "401": {
                        "description": "Неавторизован",
                        "schema": {
                            "$ref": "#/definitions/utils.Err"
                        }
                    },
                    "403": {
                        "description": "Доступ запрещен",
                        "schema": {
                            "$ref": "#/definitions/utils.Err"
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/utils.Err"
                        }
                    }
                }
            }
        },
        "/partners/{partnerID}/stat/{promoID}": {
            "get": {
                "description": "Stats for partner for specific promo",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Stats"
                ],
                "summary": "Stats endpoint",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Partner ID",
                        "name": "partnerID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Promo ID",
                        "name": "promoID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.StatPromo"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.Err"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/utils.Err"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.Err"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.ErrorResponse": {
            "description": "Структура ошибки, возвращаемая API в формате { \"error\": \"message\" }",
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Сообщение об ошибке"
                }
            }
        },
...
        "models.Stat": {
            "description": "Структура, содержащая основные бизнес-метрики использования промо-кодов.",
            "type": "object",
            "properties": {
                "active_promos_count": {
                    "description": "Количество активных промо-акций, которые в данный момент доступны клиентам.\nПромо-акция считается активной, если у нее `is_active = true`.",
                    "type": "integer",
                    "example": 5
                },
                "active_users_count": {
                    "description": "Количество уникальных пользователей, которые использовали промо-коды.\nКаждый клиент считается только один раз, независимо от количества активаций.",
                    "type": "integer",
                    "example": 85
                },
                "buy_promos_count": {
                    "description": "Общее количество покупок, совершенных с использованием промо-кодов.\nСчитается суммарное количество активаций промо, приведших к покупке.",
                    "type": "integer",
                    "example": 120
                },
                "conversion_rate": {
                    "description": "Конверсия пользователей в использование промо-кодов (в процентах).\nРассчитывается как `(ActiveUsers / TotalClients) * 100`.",
                    "type": "number",
                    "example": 17
                },
                "daily_active_users_count": {
                    "description": "Количество уникальных пользователей, активировавших промо-коды за последние 24 часа.\nПоказывает ежедневную вовлеченность клиентов в акции.",
                    "type": "integer",
                    "example": 25
                },
                "daily_buy_promo_count": {
                    "description": "Количество покупок, совершенных с промо-кодами за последние 24 часа.\nПолезно для оценки текущей активности пользователей.",
                    "type": "integer",
                    "example": 30
                },
                "limit_promos_count": {
                    "description": "Количество промо-акций, достигших лимита использования.\nНапример, если промо-код имеет ограничение `usage_limit = 100`, то он больше не учитывается после достижения лимита.",
                    "type": "integer",
                    "example": 10
                },
                "retention_30d": {
                    "description": "Количество клиентов, которые использовали промо-коды повторно в последние 30 дней.\nЭтот показатель помогает измерить уровень удержания клиентов.",
                    "type": "integer",
                    "example": 40
                },
                "top_users": {
                    "description": "Количество топ-клиентов, которые использовали промо-коды чаще всего.\nЭтот показатель можно использовать для определения лояльных клиентов.",
                    "type": "integer",
                    "example": 10
                },
                "total_users": {
                    "description": "Общее количество зарегистрированных клиентов у партнера.\nПолезно для оценки охвата бизнеса и анализа конверсии в использование промо-кодов.",
                    "type": "integer",
                    "example": 500
                }
            }
        },
        "models.StatPromo": {
            "type": "object",
            "properties": {
                "buy_promos_count": {
                    "type": "integer"
                },
                "daily_buy_promo_count": {
                    "type": "integer"
                },
                "daily_users_count": {
                    "type": "integer"
                },
                "users_count": {
                    "type": "integer"
                }
            }
        },
        "utils.Err": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}